% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.ISO3.R
\name{to.ISO3}
\alias{to.ISO3}
\title{Convert Country Names for PolarCAP Retrieval}
\usage{
to.ISO3(countries)
}
\arguments{
\item{countries}{a character vector of countries to be checked (case-insensitive).}
}
\value{
a character vector of length equal to that of \code{countries}.
}
\description{
Checks if requested countries are formatted correctly for PolarCAP and attempts to convert them to
ISO 3166-1 alpha-3 country codes if not.
}
\examples{
to.ISO3(c("ALB", "aus", "united states"))

}
